refactor : space 도메인 dto 패키지 변경
