+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/membership/controller/ApiV1InviteController.java
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.*;
 import org.tuna.zoopzoop.backend.domain.member.entity.Member;
 import org.tuna.zoopzoop.backend.domain.space.membership.entity.Membership;
 import org.tuna.zoopzoop.backend.domain.space.membership.service.MembershipService;
-import org.tuna.zoopzoop.backend.domain.space.space.dto.ResBodyForSpaceInviteList;
-import org.tuna.zoopzoop.backend.domain.space.space.dto.ResBodyForSpaceSave;
-import org.tuna.zoopzoop.backend.domain.space.space.dto.SpaceMembershipInfoWithoutAuthority;
+import org.tuna.zoopzoop.backend.domain.space.space.dto.res.ResBodyForSpaceInviteList;
+import org.tuna.zoopzoop.backend.domain.space.space.dto.res.ResBodyForSpaceSave;
+import org.tuna.zoopzoop.backend.domain.space.space.dto.etc.SpaceMembershipInfoWithoutAuthority;
 import org.tuna.zoopzoop.backend.domain.space.space.service.SpaceService;
 import org.tuna.zoopzoop.backend.global.rsData.RsData;
 import org.tuna.zoopzoop.backend.global.security.jwt.CustomUserDetails;
