+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/space/controller/ApiV1SpaceController.java
@@ -11,10 +11,10 @@ import org.tuna.zoopzoop.backend.domain.space.membership.entity.Membership;
 import org.tuna.zoopzoop.backend.domain.space.membership.enums.Authority;
 import org.tuna.zoopzoop.backend.domain.space.membership.enums.JoinState;
 import org.tuna.zoopzoop.backend.domain.space.membership.service.MembershipService;
-import org.tuna.zoopzoop.backend.domain.space.space.dto.ReqBodyForSpaceSave;
-import org.tuna.zoopzoop.backend.domain.space.space.dto.ResBodyForSpaceList;
-import org.tuna.zoopzoop.backend.domain.space.space.dto.SpaceMembershipInfo;
-import org.tuna.zoopzoop.backend.domain.space.space.dto.ResBodyForSpaceSave;
+import org.tuna.zoopzoop.backend.domain.space.space.dto.req.ReqBodyForSpaceSave;
+import org.tuna.zoopzoop.backend.domain.space.space.dto.res.ResBodyForSpaceList;
+import org.tuna.zoopzoop.backend.domain.space.space.dto.etc.SpaceMembershipInfo;
+import org.tuna.zoopzoop.backend.domain.space.space.dto.res.ResBodyForSpaceSave;
 import org.tuna.zoopzoop.backend.domain.space.space.entity.Space;
 import org.tuna.zoopzoop.backend.domain.space.space.service.SpaceService;
 import org.tuna.zoopzoop.backend.global.rsData.RsData;
