--- src/main/java/org/tuna/zoopzoop/backend/domain/space/space/dto/ResBodyForSpaceInviteList.java
@@ -1,8 +1,0 @@
-package org.tuna.zoopzoop.backend.domain.space.space.dto;
-
-import java.util.List;
-
-public record ResBodyForSpaceInviteList(
-        List<SpaceMembershipInfoWithoutAuthority> spaces
-) {
-}
