--- src/main/java/org/tuna/zoopzoop/backend/domain/space/space/dto/ResBodyForSpaceList.java
@@ -1,8 +1,0 @@
-package org.tuna.zoopzoop.backend.domain.space.space.dto;
-
-import java.util.List;
-
-public record ResBodyForSpaceList(
-        List<SpaceMembershipInfo> spaces
-) {
-}
