+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/space/dto/etc/SpaceMembershipInfo.java
@@ -1,4 +1,4 @@
-package org.tuna.zoopzoop.backend.domain.space.space.dto;
+package org.tuna.zoopzoop.backend.domain.space.space.dto.etc;
 
 import org.tuna.zoopzoop.backend.domain.space.membership.enums.Authority;
