+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/space/dto/etc/SpaceMembershipInfoWithoutAuthority.java
@@ -1,4 +1,4 @@
-package org.tuna.zoopzoop.backend.domain.space.space.dto;
+package org.tuna.zoopzoop.backend.domain.space.space.dto.etc;
 
 public record SpaceMembershipInfoWithoutAuthority(
         Integer id,
