+++ src/main/java/org/tuna/zoopzoop/backend/domain/space/space/dto/res/ResBodyForSpaceSave.java
@@ -1,4 +1,4 @@
-package org.tuna.zoopzoop.backend.domain.space.space.dto;
+package org.tuna.zoopzoop.backend.domain.space.space.dto.res;
 
 public record ResBodyForSpaceSave(
         Integer id,
